refactor(health-check): centralize health check logic into shared HealthChecker class
- Introduce HealthChecker class in shared package to encapsulate health check logic for Redis and database.
- Replace duplicated health check code in api and gateway apps with usage of HealthChecker.
- HealthChecker supports timeout, skipping checks, and error logging.
- Health check responses standardized with status codes and messages.
- Added HealthChecker exports to shared package index.
- Updated dependencies and imports to use shared HealthChecker.

This refactor improves code reuse, consistency, and maintainability of health check functionality across services.
