+++ packages/shared/src/index.ts
@@ -3,3 +3,11 @@ export {
 	type FeeBreakdown,
 	type FeeCalculationInput,
 } from "./fees.js";
+
+export {
+	HealthChecker,
+	type HealthCheckResult,
+	type HealthCheckOptions,
+	type HealthCheckDependencies,
+	type HealthResponse,
+} from "./health-check.js";
